Closeish backTest with Python Backtester

+++ WobBT/src/Cerebro.cpp
@@ -1,9 +1,9 @@
 #include "Cerebro.h"
 
-float Cerebro::run()
+double Cerebro::run()
 {
 	// Calculate Return based on buy and sell locations - commisons
-	float ret = 0;
+	double ret = 0;
 	m_Strategy->init();
 	ret = m_Strategy->run();
 	//Create a Analyzed class to return
